text infinite scroll finished!
Finally I used a decision I managed to find on CodePen. The thing is
that you have to make three blocks of the same object and then start
scrolling them with translateX, but no 100% scroll, only 69.666%. This
allows you seamlessly start your animation from the beginning when the
3rd block is actually only on its halfway and is at the same place from
which the 1st block starts moving.
However Im not finished yet. Im going to reduce the amount of blocks
and leave only two of them. Moreover, I want to start the animation
from the right border, not from the center.